FE-01: refactor code
